Tests modified for AaveV2StrollOut strategy.
A new strategy which tops-up a user's supertoken balance using Aave aTokens has been written and tested with 100% coverage
achieved. The tests written for this new strategy is same as for ERC20StrollOut strategy with many modifications. These
modifications are necessary to interact with the Aave V2 contracts in a mocked environment. The tests in ERC20StrollOut
strategy also seem to be a bit wrong though I may be wrong about that.