Reorganisation and clean-up performed.
1. There were some functions which were common to both the strategies and will probably be common to all the future strategies as
well. These functions have been moved to an abstract contract called `StrategyBase`.

2. Support for supertoken check statement has been removed from the `topUp` functions and instead put in the `checkTopUpByIndex`
function of `StrollManager`. `checkTopUpByIndex` is a view function which will be called by external keepers to check if a topup
is necessary. This will save us gas cost as earlier, this check was being performed in the `topUp` function which isn't a view
function.

3. `require` statements have been converted to `error` statements to reduce gas costs and for more informative revert statements.
Some `error` statements now contain parameters which provide more information regarding revert reasons.

4. Tests have been modified to work with the custom errors. However, instead of checking the custom error that's being thrown,
we are now checking if the function is reverting or not i.e., without any specific reason.

5. NatSpec comments have been added wherever necessary (in the interfaces). However, the same is remaining to be done in the
`IStrollManager` interface. @harsh132 please add the same from your side with necessary comments in the `StrollManager` contract as well.

Further changes to `StrollManager` must be avoided. If any feature needs to be added, note down the same and we will add it in
the next version of the contracts.