Fixed issue 12 and modified tests.
While fixing this issue, we found that the due to tests' inter-dependencies, some tests were failing. My advice would be to take a clean-state approach such that the balances of the users are predictable (same) before each test. However, for now, the tests do pass. Another thing to note is that the flow rate used in tests are very large. Although the tests work as expected, it would be better to consider smaller flow rates. Also replaced `MIN_LOWER` and `MIN_UPPER` with the values in seconds i.e., previously `MIN_LOWER=2` now it's `MIN_LOWER=helper.getSeconds(2)` indicating that `MIN_LOWER` is number of seconds in 2 days.