Added support for partial amount top-ups (issue 32).
This commit is for fixing the issue #32. However, in order to make the code work, tests had to be modified and I highly recommend to go over the modified tests once to see if the tests are as expected. The strategy contracts have been modified and although they are not breaking changes as far as keepers or `StrollManager` contract is concerned, the changes might have to be reviewed by the auditors.